Elimine app blog y deje solo appPrincipal
