--- apps/blog/models.py
@@ -1,5 +1,0 @@
-from django.db import models
-
-# Create your models here.
-class Category(models.Model):
-    name = models.CharField(max_length=20)
