+++ django_site/settings/local.py
@@ -8,7 +8,7 @@ DATABASES = {
         'ENGINE': 'django.db.backends.mysql',
         'NAME': 'django_site',
         'USER': 'root',
-        'PASSWORD':'Alexix31',
+        'PASSWORD':'nico123',
         'HOST':'',
         'PORT':'3306',
